Respect close setting

+++ lib/markdown-preview-auto-open.coffee
@@ -9,7 +9,7 @@ module.exports = MarkdownPreviewAutoOpen =
         type: 'string'
     closePreviewWhenClosingFile:
       type: 'boolean'
-      default: false
+      default: true
 
   activate: (state) ->
     process.nextTick =>
@@ -41,9 +41,10 @@ module.exports = MarkdownPreviewAutoOpen =
       workspaceView = atom.views.getView(atom.workspace)
       atom.commands.dispatch workspaceView, 'markdown-preview:toggle'
 
-    if event.item.onDidDestroy
-      event.item.onDidDestroy ->
-        for pane in atom.workspace.getPanes()
-          for item in pane.items when item.getURI() is "markdown-preview://#{event.uri}"
-            pane.destroyItem(item)
-            break
+    if atom.config.get('markdown-preview-opener.closePreviewWhenClosingFile')
+      if event.item.onDidDestroy
+        event.item.onDidDestroy ->
+          for pane in atom.workspace.getPanes()
+            for item in pane.items when item.getURI() is "markdown-preview://#{event.uri}"
+              pane.destroyItem(item)
+              break
